Start breaking dominion down into smaller models and files.
