+++ commands/base_commands/staff_commands.py
@@ -19,7 +19,8 @@ from server.utils.prettytable import PrettyTable
 from server.utils.exceptions import CommandError
 from commands.base import ArxCommand, ArxPlayerCommand
 from web.character.models import Clue, SearchTag, Revelation, StoryEmit, Flashback, CluePlotInvolvement
-from world.dominion.models import Organization, RPEvent, Propriety, AssetOwner, Plot, PlotAction, PrestigeCategory
+from world.dominion.models import Organization, RPEvent, Propriety, AssetOwner, PrestigeCategory
+from world.dominion.plots.models import Plot, PlotAction
 from typeclasses.characters import Character
 
 PERMISSION_HIERARCHY = [p.lower() for p in settings.PERMISSION_HIERARCHY]
