+++ commands/cmdsets/standard.py
@@ -144,7 +144,7 @@ try:
 except Exception as err:
     print("<<ERROR>>: Error encountered in exploration commands: %s" % err)
 try:
-    from world.dominion import plot_commands
+    from world.dominion.plots import plot_commands
 except Exception as err:
     print("<<ERROR>>: Error encountered in plot commands: %s" % err)
 try:
